extract cover.js from server.js
also adjust to new cover decision, see iobroker/iobroker.docs#273
